+++ src/app/modules/shared/components/attribute-line/attribute-line.component.ts
@@ -10,6 +10,8 @@ import { BuildStorageService } from '../../services/build-storage.service';
 })
 export class AttributeLineComponent implements OnInit, OnChanges {
 
+    public readonly MAX_RANK = 75;
+
     @Input()
     public readonly attribute: AttributeTraits | null = null;
 
@@ -51,17 +53,17 @@ export class AttributeLineComponent implements OnInit, OnChanges {
     private updateCursor() {
         if (this.attribute !== null) {
             if (this.cursorIndex === null) {
-                this.cappedCursor = this.attribute.rank;
+                this.cappedCursor = this.attribute.baseRank;
                 this.highlightRange = { min: 0, max: 0 };
-                this.unlockedRange = { min: 0, max: this.attribute.rank };
-                this.bonusRange = { min: this.attribute.rank, max: this.attribute.rank + this.attribute.bonusRank };
+                this.unlockedRange = { min: 0, max: this.attribute.baseRank };
+                this.bonusRange = { min: this.attribute.baseRank, max: this.attribute.baseRank + this.attribute.bonusRank };
             } else {
-                this.cappedCursor = Math.min(this.attribute.rank + this.remainingPoints, this.cursorIndex + 1);
-                this.bonusRange = { min: Math.max(this.cappedCursor, this.attribute.rank), max: this.attribute.rank + this.attribute.bonusRank };
-                if (this.cursorIndex < this.attribute.rank) {
-                    this.highlightRange = { min: this.cursorIndex + 1, max: Math.max(this.attribute.rank, this.cursorIndex + 1) };
+                this.cappedCursor = Math.min(this.attribute.baseRank + this.remainingPoints, this.cursorIndex + 1);
+                this.bonusRange = { min: Math.max(this.cappedCursor, this.attribute.baseRank), max: this.attribute.baseRank + this.attribute.bonusRank };
+                if (this.cursorIndex < this.attribute.baseRank) {
+                    this.highlightRange = { min: this.cursorIndex + 1, max: Math.max(this.attribute.baseRank, this.cursorIndex + 1) };
                 } else {
-                    this.highlightRange = { min: this.attribute.rank, max: this.cappedCursor };
+                    this.highlightRange = { min: this.attribute.baseRank, max: this.cappedCursor };
                 }
             }
         }
@@ -78,8 +80,8 @@ export class AttributeLineComponent implements OnInit, OnChanges {
     }
 
     private setRank(rank: number) {
-        if (this.attribute !== null && this.attribute.rank !== rank && !this.readonly) {
-            this.attribute.rank = rank;
+        if (this.attribute !== null && this.attribute.baseRank !== rank && !this.readonly) {
+            this.attribute.baseRank = rank;
             this.slormancerAttributeService.updateAttributeTraits(this.attribute);
             this.buildStorageService.saveLayer();
             this.updateCursor();
@@ -106,13 +108,14 @@ export class AttributeLineComponent implements OnInit, OnChanges {
 
     public addPoint(event: MouseEvent) {
         if (this.attribute !== null && this.remainingPoints > 0) {
-            this.setRank(this.attribute.rank + Math.min(this.remainingPoints, event.shiftKey ? 10 : 1));
+            let maxPoints = Math.min(this.remainingPoints, (this.MAX_RANK - this.attribute.baseRank))
+            this.setRank(this.attribute.baseRank + Math.min(maxPoints, event.shiftKey ? 10 : 1));
         }
     }
 
     public removePoint(event: MouseEvent) {
-        if (this.attribute !== null && this.attribute.rank > 0) {
-            this.setRank(this.attribute.rank - Math.min(this.attribute.rank, event.shiftKey ? 10 : 1));
+        if (this.attribute !== null && this.attribute.baseRank > 0) {
+            this.setRank(this.attribute.baseRank - Math.min(this.attribute.baseRank, event.shiftKey ? 10 : 1));
         }
     }