+++ src/app/modules/shared/constants.ts
@@ -23,4 +23,5 @@ export const APP_TO_GAME_VERSION_MAPPER: { [key: string]: string } = {
     '0.3.0': '0.5.01e',
     '0.3.1': '0.5.01e',
     '0.4.0': '0.6.5l',
+    '0.4.1': '0.6.5n',
 }